add general ignoreliquid class and ignore liquid in links
