+++ lib/dradis/ce.rb
@@ -9,6 +9,8 @@ module Dradis
 end
 
 require 'html/ignore_liquid_in_textile_block_codes'
+require 'html/ignore_liquid'
+require 'html/ignore_liquid_in_links'
 require 'html/no_inline_code_textile_formatter'
 
 require 'html/pipeline/dradis/code_highlight_filter'
